- Extract reusable parts into common package - Fix lint problems
